requirements.txt
asgiref==3.2.7
Django==3.0.4
django-cors-headers==3.2.1
django-shortuuidfield==0.1.3
djangorestframework==3.11.0
gunicorn==20.0.4
pytz==2019.3
shortuuid==1.0.1
six==1.14.0
sqlparse==0.3.1
